+++ cfdm/field.py
@@ -2774,8 +2774,8 @@ class Field(
     def nc_hdf5_chunksizes(self, todict=False):
         """Get the HDF5 chunking strategy for the data.
 
-        Deprecated at version NEXTVERSION. Use `nc_dataset_chunksizes`
-        instead.
+        Deprecated at version NEXTVERSION and is no longer
+        available. Use `nc_dataset_chunksizes` instead.
 
         .. versionadded:: (cfdm) 1.11.2.0
 
@@ -2835,8 +2835,8 @@ class Field(
     def nc_clear_hdf5_chunksizes(self, constructs=False):
         """Clear the HDF5 chunking strategy.
 
-        Deprecated at version NEXTVERSION. Use
-        `nc_clear_dataset_chunksizes` instead.
+        Deprecated at version NEXTVERSION and is no longer
+        available. Use `nc_clear_dataset_chunksizes` instead.
 
         .. versionadded:: (cfdm) 1.11.2.0
 
@@ -2910,8 +2910,8 @@ class Field(
     ):
         """Set the HDF5 chunking strategy.
 
-        Deprecated at version NEXTVERSION. Use
-        `nc_set_dataset_chunksizes` instead.
+        Deprecated at version NEXTVERSION and is no longer
+        available. Use `nc_set_dataset_chunksizes` instead.
 
         .. versionadded:: (cfdm) 1.11.2.0